Fix infinite loop on parent pom.xml with same artifactId
Root cause of the issue:

It wasn't really a circular reference, it was the
fact that Trivy needs to get the parent POM for
every dependency it analyzes, and it starts by
checking ../pom.xml for the parent before making
an HTTP call to Maven central. 99% of the time
../pom.xml will NOT be the parent POM for a
dependency. It detects whether ../pom.xml is the
parent by comparing the parent and dependency
artifactId (to fix this same infinite loop bug)
but in our case, the artifactId were both parent,
so I added a check for groupId as well.